pyproject.toml
[tool.poetry]
name = "wealth-ops"
version = "2.0.0"
description = "Automated swing trading system"
authors = ["Wealth-Ops Team"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
httpx = "^0.27"
pandas = "^2.2"
pyarrow = "^15.0"
yfinance = "^0.2"
boto3 = "^1.34"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^4.1"
mypy = "^1.8"
ruff = "^0.2"
moto = { version = "^5.0", extras = ["all"] }
boto3-stubs = { version = "^1.34", extras = ["essential"] }

[tool.poetry.group.infra.dependencies]
aws-cdk-lib = "^2.120"
constructs = "^10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-branch --cov-report=term-missing"
